+++ packages/insomnia/src/ui/components/buttons/grpc-send-button.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent } from 'react';
 
 import type { GrpcMethodType } from '../../../network/grpc/method';
 import { GrpcMethodTypeEnum } from '../../../network/grpc/method';
-import { Button, ButtonProps } from '../../insomnia-components';
+import { Button, ButtonProps } from '../themed-button';
 
 interface Props {
   running: boolean;
