+++ packages/insomnia/src/ui/components/buttons/settings-button.tsx
@@ -2,11 +2,11 @@ import React, { FunctionComponent } from 'react';
 import { useSelector } from 'react-redux';
 import styled from 'styled-components';
 
-import { CircleButton } from '../../insomnia-components';
 import { selectSettings } from '../../redux/selectors';
 import { Hotkey } from '../hotkey';
 import { showSettingsModal } from '../modals/settings-modal';
 import { SvgIcon } from '../svg-icon';
+import { CircleButton } from '../themed-button';
 import { Tooltip } from '../tooltip';
 
 const Wrapper = styled.div({
