+++ packages/insomnia/src/ui/components/dropdowns/dashboard-sort-dropdown.tsx
@@ -2,8 +2,9 @@ import React, { FC } from 'react';
 import styled from 'styled-components';
 
 import { DASHBOARD_SORT_ORDERS, DashboardSortOrder, dashboardSortOrderName } from '../../../common/constants';
-import { Button, Dropdown, DropdownItem } from '../../insomnia-components';
+import { Dropdown, DropdownItem } from '../../insomnia-components';
 import { SvgIcon } from '../svg-icon';
+import { Button } from '../themed-button';
 import { svgPlacementHack } from './dropdown-placement-hacks';
 
 interface DashboardSortDropdownProps {
