+++ packages/insomnia/src/ui/components/dropdowns/grpc-method-dropdown/grpc-method-dropdown-button.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent, useMemo } from 'react';
 import styled from 'styled-components';
 
 import { getGrpcPathSegments, getShortGrpcPath } from '../../../../common/grpc-paths';
-import { Button } from '../../../insomnia-components';
+import { Button } from '../../themed-button';
 import { Tooltip } from '../../tooltip';
 
 const FlexSpaceBetween = styled.span`
