+++ packages/insomnia/src/ui/components/list-group/unit-test-item.tsx
@@ -3,7 +3,7 @@ import React, { FunctionComponent, ReactNode } from 'react';
 import useToggle from 'react-use/lib/useToggle';
 import styled from 'styled-components';
 
-import { Button } from '../../insomnia-components/button';
+import { Button } from '../themed-button';
 import { SvgIcon } from '../svg-icon';
 import { ListGroupItem } from './list-group-item';
 import { UnitTestRequestSelector } from './unit-test-request-selector';
