+++ packages/insomnia/src/ui/components/proto-file/proto-file-list-item.tsx
@@ -1,8 +1,8 @@
 import React, { FunctionComponent, useCallback } from 'react';
 
 import type { ProtoFile } from '../../../models/proto-file';
-import { AsyncButton, Button } from '../../insomnia-components';
 import { Editable } from '../base/editable';
+import { AsyncButton, Button } from '../themed-button';
 import type {
   DeleteProtoFileHandler,
   RenameProtoFileHandler,
