+++ packages/insomnia/src/ui/components/proto-file/proto-list-item.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { ListGroupItem } from '../../insomnia-components';
+import { ListGroupItem } from '../list-group';
 
 export const ProtoListItem = styled(ListGroupItem).attrs(() => ({
   className: 'row-spaced',
