+++ packages/insomnia/src/ui/insomnia-components/index.ts
@@ -1,2 +1,1 @@
-export * from './button';
 export * from './dropdown';
